requirements.txt
fastapi==0.63.0  # version fixed due to complete re-overriding fastapi objects
uvicorn
asyncpg
databases~=0.4.3
python-dotenv~=0.18.0
pyyaml
starlette_context

pyjwt
pydantic
passlib
argon2-cffi # required for passlib.hash.argon2

aio_pika # amqp

aioboto3 # s3

python-multipart # required for fastapi UploadFile

line-bot-sdk # line